Introduce BridgeOptions deprecation in favour of SockJSBridgeOptions
BridgeOptions is removed in 4.0.

Avoids an abrupt breaking change due to https://github.com/vert-x3/vertx-web/pull/1553 and no corresponding deprecation in 3.9
