+++ vertx-web/src/main/asciidoc/dataobjects.adoc
@@ -4,7 +4,7 @@
 == BridgeOptions
 
 ++++
- Options for configuring the event bus bridge.
+ Deprecated: use link instead.
 ++++
 '''
 
@@ -52,6 +52,28 @@
 |[[addressRegex]]`@addressRegex`|`String`|-
 |[[match]]`@match`|`Json object`|-
 |[[requiredAuthority]]`@requiredAuthority`|`String`|-
+|===
+
+[[SockJSBridgeOptions]]
+== SockJSBridgeOptions
+
+++++
+ Options for configuring the event bus bridge.
+++++
+'''
+
+[cols=">25%,25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[inboundPermitted]]`@inboundPermitted`|`Array of link:dataobjects.html#PermittedOptions[PermittedOptions]`|-
+|[[inboundPermitteds]]`@inboundPermitteds`|`Array of link:dataobjects.html#PermittedOptions[PermittedOptions]`|-
+|[[maxAddressLength]]`@maxAddressLength`|`Number (int)`|-
+|[[maxHandlersPerSocket]]`@maxHandlersPerSocket`|`Number (int)`|-
+|[[outboundPermitted]]`@outboundPermitted`|`Array of link:dataobjects.html#PermittedOptions[PermittedOptions]`|-
+|[[outboundPermitteds]]`@outboundPermitteds`|`Array of link:dataobjects.html#PermittedOptions[PermittedOptions]`|-
+|[[pingTimeout]]`@pingTimeout`|`Number (long)`|-
+|[[replyTimeout]]`@replyTimeout`|`Number (long)`|-
 |===
 
 [[SockJSHandlerOptions]]
