+++ vertx-web/src/main/java/examples/WebExamples.java
@@ -22,7 +22,7 @@ import io.vertx.ext.bridge.BridgeEventType;
 import io.vertx.ext.bridge.PermittedOptions;
 import io.vertx.ext.web.*;
 import io.vertx.ext.web.handler.*;
-import io.vertx.ext.web.handler.sockjs.BridgeOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSBridgeOptions;
 import io.vertx.ext.web.handler.sockjs.SockJSHandler;
 import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
 import io.vertx.ext.web.sstore.ClusteredSessionStore;
@@ -894,7 +894,7 @@ public class WebExamples {
     Router router = Router.router(vertx);
 
     SockJSHandler sockJSHandler = SockJSHandler.create(vertx);
-    BridgeOptions options = new BridgeOptions();
+    SockJSBridgeOptions options = new SockJSBridgeOptions();
     // mount the bridge on the router
     router.mountSubRouter("/eventbus", sockJSHandler.bridge(options));
   }
@@ -928,7 +928,7 @@ public class WebExamples {
     PermittedOptions outboundPermitted2 = new PermittedOptions().setAddressRegex("news\\..+");
 
     // Let's define what we're going to allow from client -> server
-    BridgeOptions options = new BridgeOptions().
+    SockJSBridgeOptions options = new SockJSBridgeOptions().
       addInboundPermitted(inboundPermitted1).
       addInboundPermitted(inboundPermitted1).
       addInboundPermitted(inboundPermitted3).
@@ -947,7 +947,7 @@ public class WebExamples {
     // But only if the user is logged in and has the authority "place_orders"
     inboundPermitted.setRequiredAuthority("place_orders");
 
-    BridgeOptions options = new BridgeOptions().addInboundPermitted(inboundPermitted);
+    SockJSBridgeOptions options = new SockJSBridgeOptions().addInboundPermitted(inboundPermitted);
   }
 
   public void example48(Vertx vertx, AuthProvider authProvider) {
@@ -973,7 +973,7 @@ public class WebExamples {
     // mount the bridge on the router
     router.mountSubRouter(
       "/eventbus",
-      sockJSHandler.bridge(new BridgeOptions().addInboundPermitted(inboundPermitted)));
+      sockJSHandler.bridge(new SockJSBridgeOptions().addInboundPermitted(inboundPermitted)));
   }
 
   public void example48_1(Vertx vertx) {
@@ -984,7 +984,7 @@ public class WebExamples {
     PermittedOptions inboundPermitted = new PermittedOptions().setAddress("demo.orderService");
 
     SockJSHandler sockJSHandler = SockJSHandler.create(vertx);
-    BridgeOptions options = new BridgeOptions().addInboundPermitted(inboundPermitted);
+    SockJSBridgeOptions options = new SockJSBridgeOptions().addInboundPermitted(inboundPermitted);
 
     // mount the bridge on the router
     router.mountSubRouter(
@@ -1009,7 +1009,7 @@ public class WebExamples {
     PermittedOptions inboundPermitted = new PermittedOptions().setAddress("demo.someService");
 
     SockJSHandler sockJSHandler = SockJSHandler.create(vertx);
-    BridgeOptions options = new BridgeOptions().addInboundPermitted(inboundPermitted);
+    SockJSBridgeOptions options = new SockJSBridgeOptions().addInboundPermitted(inboundPermitted);
 
     // mount the bridge on the router
     router.mountSubRouter("/eventbus", sockJSHandler.bridge(options, be -> {
@@ -1029,7 +1029,7 @@ public class WebExamples {
 
     // Initialize SockJS handler
     SockJSHandler sockJSHandler = SockJSHandler.create(vertx);
-    BridgeOptions options = new BridgeOptions().addInboundPermitted(inboundPermitted).setPingTimeout(5000);
+    SockJSBridgeOptions options = new SockJSBridgeOptions().addInboundPermitted(inboundPermitted).setPingTimeout(5000);
 
     // mount the bridge on the router
     router.mountSubRouter("/eventbus", sockJSHandler.bridge(options, be -> {
