+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sockjs/impl/SockJSHandlerImpl.java
@@ -41,12 +41,8 @@ import io.vertx.core.logging.LoggerFactory;
 import io.vertx.core.shareddata.LocalMap;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.handler.sockjs.BridgeEvent;
-import io.vertx.ext.web.handler.sockjs.BridgeOptions;
-import io.vertx.ext.web.handler.sockjs.SockJSHandler;
-import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
-import io.vertx.ext.web.handler.sockjs.SockJSSocket;
-import io.vertx.ext.web.handler.sockjs.Transport;
+import io.vertx.ext.web.handler.sockjs.*;
+import io.vertx.ext.web.handler.sockjs.SockJSBridgeOptions;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
@@ -92,12 +88,12 @@ public class SockJSHandlerImpl implements SockJSHandler {
   }
 
   @Override
-  public Router bridge(BridgeOptions bridgeOptions) {
+  public Router bridge(SockJSBridgeOptions bridgeOptions) {
     return bridge(bridgeOptions, null);
   }
 
   @Override
-  public Router bridge(BridgeOptions bridgeOptions, Handler<BridgeEvent> bridgeEventHandler) {
+  public Router bridge(SockJSBridgeOptions bridgeOptions, Handler<BridgeEvent> bridgeEventHandler) {
     return socketHandler(new EventBusBridgeImpl(vertx, bridgeOptions, bridgeEventHandler));
   }
